refactor: move `SEMVER_REGEX` to schema 1.0
