+++ schema/1.0/validator.js
@@ -16,7 +16,7 @@ const path = require('path');
 const _ = require('lodash');
 const colors = require('colors');
 const Table = require('cli-table3');
-const { SEMVER_REGEX } = require('../../constants');
+const { SEMVER_REGEX } = require('./constants/semver-regex');
 const ValidatorError = require('../../validator-error');
 const BaseSchema = require('./base-schema');
 const PendingApproval = require('./steps/pending-approval');
